+++ config.js
@@ -1,3 +1,7 @@
+
+// Set the varibles for development environment
+require("dotenv").config();
+
 export const firebaseConnection = {
   apiKey: process.env.API_KEY,
   authDomain: process.env.AUTH_DOMAIN,
@@ -7,6 +11,8 @@ export const firebaseConnection = {
   messagingSenderId: process.env.MENSSAGING_SENDER_ID
 };
 
-export const SuperSecretDiscordToken = {
+export const superSecretDiscordToken = {
   token: process.env.DISCORD_TOKEN
 };
+
+export const projectVersion = require('./package.json').version;