+++ src/utils/commandHandler.ts
@@ -18,6 +18,7 @@ import { buy } from "../commands/buy";
 import { help } from "../commands/help";
 import { inventory } from "../commands/inventory";
 import { sell } from "../commands/sell";
+import { hello } from "../commands/hello";
 
 /**
  * Receives a message, treating it and sending to the right method
@@ -47,7 +48,6 @@ export function commandHandler(msg: Discord.Message) {
   else if (command === "help") help(msg);
   else if (command === "inventory") inventory(msg);
   else if (command === "sell") sell(msg, args[1]);
-  else {
-    msg.reply(" I didn't understand what do you mean");
-  }
+  else if (command === "hi" || command === "hello" || command === "greetings") hello(msg);
+  else msg.reply(" I didn't understand what do you mean");
 }